refactor:express server code refactoring
